RecyclerView adapter fixes
According to the MSDN documentation NewStartIndex and OldStartIndex
can be -1 for certain operations.  This is stupid.  When this happens
we need to behave as if a Reset occurred.

Fixes an issue with Replace where we need to use the old index/items.
Fixes an issue with Move where if one item was moved (the default with
ObservableCollection) OldItems and NewItems were equal so the new/old
positions were equivalent.

Fixes #310